requirements.txt
python-telegram-bot==20.7
APScheduler==3.10.4
python-dotenv==1.0.0
psycopg2-binary==2.9.9
flask==2.3.0
